% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare.data.R
\docType{data}
\name{data.dtm}
\alias{data.dtm}
\title{Document-term matrix for adjectives in TripAdvisor hotel reviews}
\format{An object of class \code{dgCMatrix} with 500 rows and 200 columns.}
\usage{
data.dtm
}
\description{
A 500-by-200 document-term matrix for 200 adjectives appearing in 500 TripAdvisor reviews.
The document-term matrix is in sparse format.
}
\seealso{
\code{\link{data.rating}}, \code{\link{data.hc}}.
}
\keyword{datasets}
