% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare.data.R
\docType{data}
\name{data.hc}
\alias{data.hc}
\title{Hierarchical clustering tree for adjectives in TripAdvisor data set}
\format{An object of class \code{hclust} of length 7.}
\source{
Embeddings available at \url{http://nlp.stanford.edu/data/glove.6B.zip}
}
\usage{
data.hc
}
\description{
An \code{hclust} tree for the 200 adjectives appearing in the TripAdvisor reviews.
The tree was generated with 100-dimensional word embeddings pre-trained by GloVe
(Pennington et al., 2014) on Gigaword5 and Wikipedia2014 corpora for the adjectives.
}
\references{
Pennington, J., Socher, R., and Manning, C. D. (2014).
Glove: Global vectors for word representation.
\emph{In Empirical Methods in Natural Language Processing (EMNLP)}, pages 1532–1543.
}
\keyword{datasets}
