% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare.fit.R
\name{plot.rarefit}
\alias{plot.rarefit}
\title{Plot a rarefit object}
\usage{
\method{plot}{rarefit}(x, y, alpha_index = NULL, type = c("beta", "gamma"), ...)
}
\arguments{
\item{x}{rarefit object}

\item{y}{unused argument}

\item{alpha_index}{specify which alpha value to plot}

\item{type}{"beta" or "gamma" determines which path is to be plotted}

\item{...}{additional parameters passed to `matplot`}
}
\description{
Plot a rarefit object
}
