% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare.fit.R
\name{print.rarefit}
\alias{print.rarefit}
\title{Print information about a rarefit object}
\usage{
\method{print}{rarefit}(x, ...)
}
\arguments{
\item{x}{rarefit object}

\item{...}{additional parameters (not used)}
}
\description{
Print information about a rarefit object
}
