% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rare.data.R
\docType{data}
\name{data.rating}
\alias{data.rating}
\title{TripAdvisor hotel review ratings}
\format{An object of class \code{integer} of length 500.}
\source{
TripAdvisor Data Set used in \url{https://www.cs.virginia.edu/~hw5x/paper/rp166f-wang.pdf}
}
\usage{
data.rating
}
\description{
A length-500 TripAdvisor review ratings on the scale 1 to 5.
}
\keyword{datasets}
